+++ static/js/redux/constants/actionTypes.jsx
@@ -32,8 +32,3 @@ export const INIT_STATE = "INIT_STATE";
 // NOTIFICATIONS
 export const TOKEN_REGISTERED = "TOKEN_REGISTERED";
 export const UNREGISTER_TOKEN = "UNREGISTER_TOKEN";
-
-// FIND FRIENDS & PEER MODAL
-export const TOGGLE_PEER_MODAL = "TOGGLE_PEER_MODAL";
-export const PEER_MODAL_LOADING = "PEER_MODAL_LOADING";
-export const PEER_MODAL_LOADED = "PEER_MODAL_LOADED";
